+++ Terra/src/utopia/terra/controller/coordinate/world/CircleOfEarth.scala
@@ -5,7 +5,7 @@ import utopia.paradigm.shape.shape2d.vector.Vector2D
 import utopia.paradigm.shape.shape3d.Vector3D
 import utopia.terra.controller.coordinate.GlobeMath
 import utopia.terra.model.angular.LatLong
-import utopia.terra.model.enumeration.CompassDirection.{North, South}
+import utopia.terra.model.enumeration.CompassDirection.South
 import utopia.terra.model.world.circle.{CirclePoint, CircleSurfacePoint}
 
 /**
@@ -30,8 +30,14 @@ object CircleOfEarth extends WorldView[Vector2D, Vector3D, CircleSurfacePoint, C
 	  * in vector coordinate system.
 	  */
 	val equatorVectorRadius = 100000.0
+	/**
+	  * Radius of the equator circle in this world view,
+	  * in "distance"
+	units
+	*/
+	val equatorCircleRadius = GlobeMath.earthRadiusAtEquator * math.Pi / 2.0
 	
-	override val unitDistance = GlobeMath.earthRadiusAtEquator / equatorVectorRadius
+	override val unitDistance = equatorCircleRadius / equatorVectorRadius
 	
 	
 	// IMPLEMENTED  -------------------------
@@ -42,29 +48,30 @@ object CircleOfEarth extends WorldView[Vector2D, Vector3D, CircleSurfacePoint, C
 	override def surfaceVector(vector: Vector2D): CircleSurfacePoint = CircleSurfacePoint(vector)
 	override def aerialVector(vector: Vector3D): CirclePoint = CirclePoint(vector)
 	
-	
-	// OTHER    -----------------------------
-	
 	/**
-	 * @param latLong A latitude-longitude coordinate
-	 * @return A 2D vector representation of that coordinate along the circular plane
-	 */
-	def latLongToVector(latLong: LatLong) = {
+	  * @param latLong A latitude-longitude coordinate
+	  * @return A 2D vector representation of that coordinate along the circular plane
+	  */
+	override def latLongToVector(latLong: LatLong) = {
 		// Distance is 0 at the north pole (-90 latitude),
 		// 100 000 at the equator (0.0 latitude) and
 		// 200 000 at the southern rim (90 latitude)
-		val distance = (latLong(South) - North.radians(math.Pi)).clockwiseRadians * equatorVectorRadius * 2.0 / math.Pi
+		// V = (lat + 90 degrees) * R / 90 degrees
+		// 90 degrees = Pi/2
+		val vectorDistance = (latLong + South.radians(math.Pi / 2.0))(South).clockwiseRadians *
+			equatorVectorRadius * 2.0 / math.Pi
 		// Direction matches the longitude, because 0 angle (right) matches the longitude of 0 (Greenwich, England)
-		Vector2D.lenDir(distance, latLong.longitude)
+		Vector2D.lenDir(vectorDistance, latLong.longitude)
 	}
 	/**
-	 * @param vector A vector relative to the circle origin along the X-Y plane
-	 *               (where X axis runs from the north pole towards Greenwich, England)
-	 * @return A latitude-longitude coordinate that matches that vector position
-	 */
-	def vectorToLatLong(vector: Vector2D) = {
+	  * @param vector A vector relative to the circle origin along the X-Y plane
+	  *               (where X axis runs from the north pole towards Greenwich, England)
+	  * @return A latitude-longitude coordinate that matches that vector position
+	  */
+	override def vectorToLatLong(vector: Vector2D) = {
 		// 0.0 latitude is at the equator, which lies at length 100 000
-		val latitude = South.degrees((1 - vector.length / equatorVectorRadius) * 90.0)
+		// Lat = 90 degrees * vectorLength / R - 90 degrees
+		val latitude = South.radians(math.Pi * vector.length / (equatorVectorRadius * 2.0) - (math.Pi / 2.0))
 		// Longitude matches the vector direction exactly
 		LatLong(latitude, vector.direction)
 	}